Introduce AbstractPluginBuilder base class
Extract the common state and dispatch logic of PluginClassBuilder into
an abstract base class. The base owns pluginClassName, packageName,
language, type, bindingModifiers, applyActionDeclaration, and
customApplyActionCode, plus the shared fluent setters and the build()
method that dispatches on language to a renderJava() / renderKotlin()
abstract pair.
